Ignore everything I said for the past few commits.
This should not be this hard, and I don't have time for it! Sheesh.
Config is now handled by a default config.php (in the repo) which
checks for the existence of config.inc (not in the repo) before using
default values. config.inc.example (tracked) provides template for
creating config.inc. OK. ONWARD.